SDC-16770. field type converter is not ignoring empty string when converting them to double
Added a setting to Field Type Converter so that the user can choose what to do with empty strings to avoid such records going to error records.

The user can choose to delete those fields, convert them into a null of the target type or leaving them as is.

I am returning null from convertField to signal fields that have to be deleted.

Change-Id: I1a571cc7c1a0353794a910d8c0ca163c048fbc41
Reviewed-on: https://review.streamsets.net/c/datacollector/+/40824
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Loretta Chen <loretta@streamsets.com>